python/requirements.txt
###### IF ADDDING NEW REQUIREMENTS PLEASE READ ######
# This file is used both for python and pypy setup files
# Read description of setup_pypy for pypy compatibility steps

absl-py==0.8.1
aws-requests-auth==0.4.2
awscli==1.19.105
boto3==1.17.64
botocore==1.20.105
cacheout==0.11.2
confluent-kafka==1.4.2
coverage==5.1
cryptography==3.3
distogram==1.6.0
dse-driver==2.11.0
elasticsearch==7.1.0
Flask==0.10.1
flask-compress==1.9.0
Flask-Injector==0.12.3
Flask-HTTPAuth==4.2.0
itsdangerous==2.0.1
kafka-python==1.4.7
lru-dict==1.1.6
lz4==3.1.3
moz-sql-parser==3.18.19324
mock===3.0.5
netifaces==0.10.9
numpy==1.18.1
orjson==3.5.3
pandas==1.0.4
pglast==1.14
# TODO: 0.5.9 is latest and used to work, but broke today 04/13
# switched to older version as a workaround
pydruid==0.5.8
pyformance==0.4
pynacl==1.3.0
pytest==5.3.2
pytrie==0.3.1
PyYAML==5.2
PyYAML==5.2
retrying==1.3.3
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.22.0
scipy==1.7.1
sshtunnel==0.4.0
tdigest
werkzeug==0.16.0
